+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.7.2] - Unreleased
 ### Added
 * Lookup warehouses by `name` for Terraform backend
+* Databricks `Permissions` declaration from the stack root. 
 
 ## [0.7.1] - 2025-03-20
 ### Fixed